Refactor: Wiki/ContentPage - Add: MethodInfo component - Add: ContentPageContainer component
